Replace static_cast<> with C-style casts and fix real conversion warnings

+++ src/dichotomy.cc
@@ -37,7 +37,7 @@ using namespace sjpeg;
 #define DBG_PRINT 0
 
 // convergence is considered reached if |dq| < kdQLimit.
-static const float kdQLimit = 0.15;
+static const float kdQLimit = 0.15f;
 
 static float Clamp(float v, float min, float max) {
   return (v < min) ? min : (v > max) ? max : v;
@@ -46,7 +46,7 @@ static float Clamp(float v, float min, float max) {
 bool SearchHook::Setup(const EncoderParam& param) {
   for_size = (param.target_mode == EncoderParam::TARGET_SIZE);
   target = param.target_value;
-  tolerance = param.tolerance / 100.;
+  tolerance = param.tolerance / 100.f;
   qmin = (param.qmin < 0) ? 0 : param.qmin;
   qmax = (param.qmax > 100) ? 100 :
          (param.qmax < param.qmin) ? param.qmin : param.qmax;
@@ -67,7 +67,7 @@ bool SearchHook::Update(float result) {
     qmin = q;
   }
   const float last_q = q;
-  q = (qmin + qmax) / 2.;
+  q = (qmin + qmax) / 2.f;
   done = (fabs(q - last_q) < kdQLimit);
   if (DBG_PRINT) {
     printf(" -> next-q=%.2f\n", last_q);
@@ -317,11 +317,10 @@ void Encoder::BlocksSize(int nb_mbs, const DCTCoeffs* coeffs,
       }
       const uint32_t suffix = rl[i].level_;
       const size_t nbits = suffix & 0x0f;
-      const int sym = (run << 4) | nbits;
+      const int sym = (int)((run << 4) | nbits);
       assert(nbits > 0);   // as in CodeBlock(): zero only comes from the ZRL
 #if defined(SJPEG_HAVE_64BIT)
-      bc->AddPackedCodeAndSuffix(codes[sym], suffix >> 4,
-                                 static_cast<int>(nbits));
+      bc->AddPackedCodeAndSuffix(codes[sym], suffix >> 4, (int)nbits);
 #else
       bc->AddPackedCode(codes[sym]);
       bc->AddBits(suffix >> 4, nbits);
@@ -364,15 +363,16 @@ float Encoder::ComputeSize(const DCTCoeffs* coeffs) {
 float Encoder::GetPSNR(uint64_t err, uint64_t size) {
   // This expression is written such that it gives the same result on ARM
   // and x86 (for large values of err/size in particular). Don't change it!
-  return (err > 0 && size > 0) ? 4.3429448f * log(size / (err / 255. / 255.))
-                               : 99.f;
+  return (err > 0 && size > 0)
+             ? (float)(4.3429448f *
+                       log((double)size / ((double)err / 255. / 255.)))
+             : 99.f;
 }
 
 uint64_t Encoder::ComputePSNRSlice(int y_start, int y_end) const {
   uint64_t error = 0;
-  const int16_t* in =
-      in_blocks_ + static_cast<size_t>(y_start) * mb_w_ * 64 * mcu_blocks_;
-  const size_t slice_mbs = static_cast<size_t>(y_end - y_start) * mb_w_;
+  const int16_t* in = in_blocks_ + (size_t)y_start * mb_w_ * 64 * mcu_blocks_;
+  const size_t slice_mbs = (size_t)(y_end - y_start) * mb_w_;
   for (size_t n = 0; n < slice_mbs; ++n) {
     for (int c = 0; c < nb_comps_; ++c) {
       const Quantizer* const Q = &quants_[quant_idx_[c]];
@@ -391,6 +391,6 @@ float Encoder::ComputePSNR() const {
   if (num_slices > 1) return ComputePSNRMultiThreaded(num_slices);
 #endif
   const uint64_t error = ComputePSNRSlice(0, mb_h_);
-  const size_t nb_mbs = static_cast<size_t>(mb_w_) * mb_h_;
+  const size_t nb_mbs = (size_t)mb_w_ * mb_h_;
   return GetPSNR(error, 64ull * nb_mbs * mcu_blocks_);
 }
